# fix: message templates alignment compliance