--- a/src/Serilog/Serilog.sln
+++ b/src/Serilog/Serilog.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.5.002.0
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Serilog", "Serilog.csproj", "{EE48D89D-E39C-439A-AB62-3026C0FC322C}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{EE48D89D-E39C-439A-AB62-3026C0FC322C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{EE48D89D-E39C-439A-AB62-3026C0FC322C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{EE48D89D-E39C-439A-AB62-3026C0FC322C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{EE48D89D-E39C-439A-AB62-3026C0FC322C}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {1B12889B-C018-4DE2-90F3-C315DBB86E88}
+	EndGlobalSection
+EndGlobal