Based on our conversation in https://github.com/messagetemplates/grammar/pull/6

This brings alignment parsing in line with the message-templates spec which specifies: `Alignment := -?[0-9]+`. I've benchmarked it locally on my laptop and desktop - these are my results from my laptop. Performance ends up the same in all current benchmarks and I'd expect that in alignment heavy workloads it'd be the sameish, maybe better

The method is a bit of a misnomer that I've introduced
`IsValidAlignment` because it really only checks that it doesn't contain any invalid characters in slots but `-` for example would pass. Not worrisome because we parse it right afterwards but might be a better name. 

I tried both benchmarks with AggressiveInlining but the compiler was smart enough to do it already so there was no perf change. Below are the benchmarks before and after. All of the memory benchmarks are approximately the same & def within the distributions noise in general while the parsing seems to be very similar as well which I'd expect given there's no alignment in the default parsing. 


## Before
### Allocations

|               Method |         Mean |     Error |    StdDev |  Ratio | RatioSD |   Gen0 |   Gen1 | Allocated | Alloc Ratio |
|--------------------- |-------------:|----------:|----------:|-------:|--------:|-------:|-------:|----------:|------------:|
|             LogEmpty |     5.135 ns | 0.0135 ns | 0.0120 ns |   1.00 |    0.00 |      - |      - |         - |          NA |
| LogEmptyWithEnricher |    18.742 ns | 0.0592 ns | 0.0554 ns |   3.65 |    0.01 |      - |      - |         - |          NA |
|            LogScalar |   212.647 ns | 0.3584 ns | 0.3352 ns |  41.42 |    0.10 | 0.0572 |      - |     360 B |          NA |
|        LogDictionary | 1,419.678 ns | 2.8334 ns | 2.6503 ns | 276.52 |    0.93 | 0.3490 |      - |    2192 B |          NA |
|          LogSequence |   330.138 ns | 0.6384 ns | 0.5972 ns |  64.29 |    0.16 | 0.0916 |      - |     576 B |          NA |
|         LogAnonymous | 1,792.173 ns | 7.1056 ns | 6.2989 ns | 349.01 |    1.44 | 0.4616 | 0.0019 |    2896 B |          NA |
|        FormatMessage |    62.933 ns | 0.2528 ns | 0.2365 ns |  12.26 |    0.05 | 0.0050 |      - |      32 B |          NA |

### Parsing

|                       Method |      Mean |    Error |   StdDev | Ratio | RatioSD |   Gen0 |   Gen1 | Allocated | Alloc Ratio |
|----------------------------- |----------:|---------:|---------:|------:|--------:|-------:|-------:|----------:|------------:|
|                EmptyTemplate |  74.78 ns | 0.488 ns | 0.456 ns |  1.00 |    0.00 | 0.0356 |      - |     224 B |        1.00 |
| DefaultConsoleOutputTemplate | 777.85 ns | 2.501 ns | 2.217 ns | 10.40 |    0.07 | 0.3443 | 0.0010 |    2160 B |        9.64 |

## After
### Allocations
| Method               |         Mean |     Error |    StdDev |  Ratio | RatioSD |   Gen0 |   Gen1 | Allocated | Alloc Ratio |
| -------------------- | -----------: | --------: | --------: | -----: | ------: | -----: | -----: | --------: | ----------: |
| LogEmpty             |     5.137 ns | 0.0126 ns | 0.0112 ns |   1.00 |    0.00 |      - |      - |         - |          NA |
| LogEmptyWithEnricher |    18.876 ns | 0.0545 ns | 0.0455 ns |   3.67 |    0.01 |      - |      - |         - |          NA |
| LogScalar            |   163.145 ns | 0.2094 ns | 0.1856 ns |  31.76 |    0.08 | 0.0572 |      - |     360 B |          NA |
| LogDictionary        | 2,168.033 ns | 3.9589 ns | 3.7032 ns | 422.02 |    1.11 | 0.3471 |      - |    2192 B |          NA |
| LogSequence          |   323.136 ns | 0.6160 ns | 0.5461 ns |  62.90 |    0.16 | 0.0916 |      - |     576 B |          NA |
| LogAnonymous         | 1,787.612 ns | 4.0751 ns | 3.8119 ns | 347.98 |    1.26 | 0.4616 | 0.0019 |    2896 B |          NA |
| FormatMessage        |    63.315 ns | 0.3750 ns | 0.3508 ns |  12.33 |    0.07 | 0.0050 |      - |      32 B |          NA |

### Parsing 
| Method                       |      Mean |    Error |   StdDev | Ratio | RatioSD |   Gen0 |   Gen1 | Allocated | Alloc Ratio |
| ---------------------------- | --------: | -------: | -------: | ----: | ------: | -----: | -----: | --------: | ----------: |
| EmptyTemplate                |  75.94 ns | 1.174 ns | 1.098 ns |  1.00 |    0.00 | 0.0356 |      - |     224 B |        1.00 |
| DefaultConsoleOutputTemplate | 778.35 ns | 2.074 ns | 1.732 ns | 10.27 |    0.15 | 0.3443 | 0.0010 |    2160 B |        9.64 |