@nblumhardt 2024-05-19
Thanks for checking this out, @Insomniak47 :+1:

It seems like the change should be reasonable and safe to make.

Is there any reason we couldn't switch to using `int.TryParse(...)` on the whole alignment value, with specific number styles and invariant culture, and then derive the direction from whether the result is negative or not? Seems like it'd be advantageous for us to get rid of the unnecessary `Substring()` call, and could be a bit faster, if we're lucky.

It looks like you may have branched off of `main`, there's a stack of additional unrelated commits in there, perhaps reapplying this from `dev` would clean it up?