@Insomniak47 2024-05-20
> Is there any reason we couldn't switch to using int.TryParse(...) on the whole alignment value, with specific number styles and invariant culture

I think we could tryparse the whole thing but I don't think that we can use the output because `-0` and `0` would end up the same. We can still just check the first character but it'd still remove the substring call which would be nice.

> It looks like you may have branched off of main, there's a stack of additional unrelated commits in there, perhaps reapplying this from dev would clean it up?

Yeah for sure, I'll rebase onto dev. I noticed this when I was putting the PR up then I got distracted 😄