@Insomniak47 2024-05-20
Alrighty, I've moved to just the parsing, I added an additional check to make sure that '+' also isn't at the beginning + a test for it. I've also rebased on dev. 

In really heavy alignment based workflows this might end up a little faster which would be neat. Looks like it allocates one additional few bytes in gen0 in the alloc benches in specific cases. I'm assuming that's because `Int32.TryParse` allocates and in the allocation benches it previously never had to hit it but it's still a small overall impact. I can take a peek at eliding it if you'd like

|               Method |         Mean |     Error |    StdDev |  Ratio | RatioSD |   Gen0 | Allocated | Alloc Ratio |
|--------------------- |-------------:|----------:|----------:|-------:|--------:|-------:|----------:|------------:|
|             LogEmpty |     5.032 ns | 0.0433 ns | 0.0362 ns |   1.00 |    0.00 |      - |         - |          NA |
| LogEmptyWithEnricher |    19.565 ns | 0.0337 ns | 0.0315 ns |   3.89 |    0.03 |      - |         - |          NA |
|            LogScalar |   228.838 ns | 0.4075 ns | 0.3812 ns |  45.49 |    0.34 | 0.1721 |     360 B |          NA |
|        LogDictionary | 2,347.531 ns | 4.0454 ns | 3.3781 ns | 466.54 |    3.34 | 1.0567 |    2216 B |          NA |
|          LogSequence |   614.448 ns | 1.2356 ns | 1.0953 ns | 122.11 |    0.96 | 0.2747 |     576 B |          NA |
|         LogAnonymous | 3,290.065 ns | 5.4689 ns | 4.5667 ns | 653.86 |    4.83 | 1.4076 |    2944 B |          NA |
|        FormatMessage |    93.986 ns | 0.4202 ns | 0.3930 ns |  18.67 |    0.16 | 0.0153 |      32 B |          NA |