@nblumhardt 2024-05-21
Looks good! Are those final After numbers from the same benchmarking setup as the Before series? Looks like some unexpected regressions, guessing it's a benchmarking issue rather than a real regression but just keen to double-check :)