@Insomniak47 2024-05-21
> Looks good! Are those final After numbers from the same benchmarking setup as the Before series? Looks like some unexpected regressions, guessing it's a benchmarking issue rather than a real regression but just keen to double-check :)
 
The last numbers I posted in https://github.com/serilog/serilog/pull/2064#issuecomment-2120664136 are against dev not main. The dev allocations numbers are the same (no regressions) and the perf is == or better I'll post an updated before/after this evening

The ones that are in there now: 
- Before is from master
- After is from dev