@Insomniak47 2024-05-22
# Before -> On dev
|               Method |         Mean |      Error |     StdDev |  Ratio | RatioSD |   Gen0 | Allocated | Alloc Ratio |
|--------------------- |-------------:|-----------:|-----------:|-------:|--------:|-------:|----------:|------------:|
|             LogEmpty |     5.054 ns |  0.0113 ns |  0.0106 ns |   1.00 |    0.00 |      - |         - |          NA |
| LogEmptyWithEnricher |    20.017 ns |  0.0267 ns |  0.0237 ns |   3.96 |    0.01 |      - |         - |          NA |
|            LogScalar |   238.454 ns |  2.4491 ns |  2.2909 ns |  47.18 |    0.44 | 0.1721 |     360 B |          NA |
|        LogDictionary | 2,426.358 ns | 36.6736 ns | 34.3045 ns | 480.10 |    6.45 | 1.0567 |    2216 B |          NA |
|          LogSequence |   618.981 ns |  5.3134 ns |  4.7102 ns | 122.50 |    1.04 | 0.2747 |     576 B |          NA |
|         LogAnonymous | 3,272.722 ns | 20.1617 ns | 16.8359 ns | 647.72 |    3.58 | 1.4076 |    2944 B |          NA |
|        FormatMessage |    95.447 ns |  0.4545 ns |  0.4251 ns |  18.89 |    0.10 | 0.0153 |      32 B |          NA |

|                       Method |       Mean |    Error |   StdDev | Ratio | RatioSD |   Gen0 | Allocated | Alloc Ratio |
|----------------------------- |-----------:|---------:|---------:|------:|--------:|-------:|----------:|------------:|
|                EmptyTemplate |   167.7 ns |  2.33 ns |  2.18 ns |  1.00 |    0.00 | 0.1070 |     224 B |        1.00 |
| DefaultConsoleOutputTemplate | 1,436.8 ns | 13.05 ns | 12.20 ns |  8.57 |    0.11 | 1.0509 |    2200 B |        9.82 |

After -> This PR

|               Method |         Mean |      Error |     StdDev |  Ratio | RatioSD |   Gen0 | Allocated | Alloc Ratio |
|--------------------- |-------------:|-----------:|-----------:|-------:|--------:|-------:|----------:|------------:|
|             LogEmpty |     5.056 ns |  0.0140 ns |  0.0125 ns |   1.00 |    0.00 |      - |         - |          NA |
| LogEmptyWithEnricher |    20.184 ns |  0.1148 ns |  0.1074 ns |   3.99 |    0.02 |      - |         - |          NA |
|            LogScalar |   236.519 ns |  1.6767 ns |  1.4864 ns |  46.78 |    0.29 | 0.1721 |     360 B |          NA |
|        LogDictionary | 2,425.448 ns | 32.1892 ns | 30.1098 ns | 480.11 |    6.00 | 1.0567 |    2216 B |          NA |
|          LogSequence |   621.591 ns |  1.9883 ns |  1.8599 ns | 122.92 |    0.52 | 0.2747 |     576 B |          NA |
|         LogAnonymous | 3,293.746 ns | 43.2289 ns | 40.4363 ns | 651.85 |    8.00 | 1.4076 |    2944 B |          NA |
|        FormatMessage |    94.241 ns |  0.7545 ns |  0.6689 ns |  18.64 |    0.12 | 0.0153 |      32 B |          NA |

|                       Method |       Mean |   Error |  StdDev | Ratio | RatioSD |   Gen0 | Allocated | Alloc Ratio |
|----------------------------- |-----------:|--------:|--------:|------:|--------:|-------:|----------:|------------:|
|                EmptyTemplate |   164.9 ns | 0.65 ns | 0.55 ns |  1.00 |    0.00 | 0.1070 |     224 B |        1.00 |
| DefaultConsoleOutputTemplate | 1,457.9 ns | 7.40 ns | 6.92 ns |  8.85 |    0.05 | 1.0509 |    2200 B |        9.82 |


So no regression in mem usage, no perf regression generally